feat: new bbox definition with Rect
