+++ docling_core/types/doc/__init__.py
@@ -5,7 +5,7 @@
 
 """Package for models defined by the Document type."""
 
-from .base import BoundingBox, CoordOrigin, ImageRefMode, Size
+from .base import BoundingBox, BoundingRectangle, CoordOrigin, ImageRefMode, Size
 from .document import (
     BaseAnnotation,
     ChartBar,
