+++ docling_core/types/doc/page.py
@@ -27,7 +27,7 @@ from PIL import ImageColor, ImageDraw, ImageFont
 from PIL.ImageFont import FreeTypeFont
 from pydantic import AnyUrl, BaseModel, Field, model_validator
 
-from docling_core.types.doc.base import BoundingBox, CoordOrigin
+from docling_core.types.doc.base import BoundingBox, BoundingRectangle, CoordOrigin, Coord2D
 from docling_core.types.doc.document import ImageRef
 
 _logger = logging.getLogger(__name__)
@@ -79,166 +79,6 @@ class ColorRGBA(BaseModel):
     def __iter__(self):
         """Yield the color components for iteration."""
         yield from (self.r, self.g, self.b, self.a)
-
-
-class Coord2D(NamedTuple):
-    """A 2D coordinate with x and y components."""
-
-    x: float
-    y: float
-
-
-class BoundingRectangle(BaseModel):
-    """Model representing a rectangular boundary with four corner points."""
-
-    r_x0: float
-    r_y0: float
-
-    r_x1: float
-    r_y1: float
-
-    r_x2: float
-    r_y2: float
-
-    r_x3: float
-    r_y3: float
-
-    coord_origin: CoordOrigin = CoordOrigin.BOTTOMLEFT
-
-    @property
-    def width(self) -> float:
-        """Calculate the width of the rectangle."""
-        return np.sqrt((self.r_x1 - self.r_x0) ** 2 + (self.r_y1 - self.r_y0) ** 2)
-
-    @property
-    def height(self) -> float:
-        """Calculate the height of the rectangle."""
-        return np.sqrt((self.r_x3 - self.r_x0) ** 2 + (self.r_y3 - self.r_y0) ** 2)
-
-    @property
-    def angle(self) -> float:
-        """Calculate the angle of the rectangle in radians (0-2pi range)."""
-        p_0 = ((self.r_x0 + self.r_x3) / 2.0, (self.r_y0 + self.r_y3) / 2.0)
-        p_1 = ((self.r_x1 + self.r_x2) / 2.0, (self.r_y1 + self.r_y2) / 2.0)
-
-        delta_x, delta_y = p_1[0] - p_0[0], p_1[1] - p_0[1]
-
-        if abs(delta_y) < 1.0e-3:
-            angle = 0.0
-        elif abs(delta_x) < 1.0e-3:
-            angle = np.pi / 2.0 if delta_y > 0 else -np.pi / 2.0
-        else:
-            angle = math.atan(delta_y / delta_x)
-        if delta_x < 0:
-            angle += np.pi
-        if angle < 0:
-            angle += 2 * np.pi
-        return angle
-
-    @property
-    def angle_360(self) -> int:
-        """Calculate the angle of the rectangle in degrees (0-360 range)."""
-        return round(self.angle / np.pi * 180)
-
-    @property
-    def centre(self):
-        """Calculate the center point of the rectangle."""
-        return (self.r_x0 + self.r_x1 + self.r_x2 + self.r_x3) / 4.0, (
-            self.r_y0 + self.r_y1 + self.r_y2 + self.r_y3
-        ) / 4.0
-
-    def to_bounding_box(self) -> BoundingBox:
-        """Convert to a BoundingBox representation."""
-        if self.coord_origin == CoordOrigin.BOTTOMLEFT:
-            top = max([self.r_y0, self.r_y1, self.r_y2, self.r_y3])
-            bottom = min([self.r_y0, self.r_y1, self.r_y2, self.r_y3])
-        else:
-            top = min([self.r_y0, self.r_y1, self.r_y2, self.r_y3])
-            bottom = max([self.r_y0, self.r_y1, self.r_y2, self.r_y3])
-
-        left = min([self.r_x0, self.r_x1, self.r_x2, self.r_x3])
-        right = max([self.r_x0, self.r_x1, self.r_x2, self.r_x3])
-
-        return BoundingBox(
-            l=left,
-            b=bottom,
-            r=right,
-            t=top,
-            coord_origin=self.coord_origin,
-        )
-
-    @classmethod
-    def from_bounding_box(cls, bbox: BoundingBox) -> "BoundingRectangle":
-        """Convert a BoundingBox into a BoundingRectangle."""
-        return cls(
-            r_x0=bbox.l,
-            r_y0=bbox.b,
-            r_x2=bbox.r,
-            r_y2=bbox.t,
-            r_x1=bbox.r,
-            r_y1=bbox.b,
-            r_x3=bbox.l,
-            r_y3=bbox.t,
-            coord_origin=bbox.coord_origin,
-        )
-
-    def to_polygon(self) -> List[Coord2D]:
-        """Convert to a list of point coordinates forming a polygon."""
-        return [
-            Coord2D(self.r_x0, self.r_y0),
-            Coord2D(self.r_x1, self.r_y1),
-            Coord2D(self.r_x2, self.r_y2),
-            Coord2D(self.r_x3, self.r_y3),
-        ]
-
-    def to_bottom_left_origin(self, page_height: float) -> "BoundingRectangle":
-        """Convert coordinates to use bottom-left origin.
-
-        Args:
-            page_height: The height of the page
-
-        Returns:
-            BoundingRectangle with bottom-left origin
-        """
-        if self.coord_origin == CoordOrigin.BOTTOMLEFT:
-            return self
-        elif self.coord_origin == CoordOrigin.TOPLEFT:
-            return BoundingRectangle(
-                r_x0=self.r_x0,
-                r_x1=self.r_x1,
-                r_x2=self.r_x2,
-                r_x3=self.r_x3,
-                r_y0=page_height - self.r_y0,
-                r_y1=page_height - self.r_y1,
-                r_y2=page_height - self.r_y2,
-                r_y3=page_height - self.r_y3,
-                coord_origin=CoordOrigin.BOTTOMLEFT,
-            )
-
-    def to_top_left_origin(self, page_height: float) -> "BoundingRectangle":
-        """Convert coordinates to use top-left origin.
-
-        Args:
-            page_height: The height of the page
-
-        Returns:
-            BoundingRectangle with top-left origin
-        """
-        if self.coord_origin == CoordOrigin.TOPLEFT:
-            return self
-        elif self.coord_origin == CoordOrigin.BOTTOMLEFT:
-            return BoundingRectangle(
-                r_x0=self.r_x0,
-                r_x1=self.r_x1,
-                r_x2=self.r_x2,
-                r_x3=self.r_x3,
-                r_y0=page_height - self.r_y0,
-                r_y1=page_height - self.r_y1,
-                r_y2=page_height - self.r_y2,
-                r_y3=page_height - self.r_y3,
-                coord_origin=CoordOrigin.TOPLEFT,
-            )
-
 
 class OrderedElement(BaseModel):
     """Base model for elements that have an ordering index."""
